inline completions
