+++ src/vs/editor/contrib/inlineCompletions/browser/inlineCompletionsAccessibleView.ts
@@ -1,0 +1,64 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Disposable } from 'vs/base/common/lifecycle';
+import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
+import { InlineCompletionContextKeys } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionContextKeys';
+import { InlineCompletionsController } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionsController';
+import { AccessibleViewType, IAccessibleViewService, AccessibleViewProviderId } from 'vs/platform/accessibility/browser/accessibleView';
+import { IAccessibleViewImplentation } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
+import { ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
+import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+
+export class InlineCompletionsAccessibleView extends Disposable implements IAccessibleViewImplentation {
+	readonly type = AccessibleViewType.View;
+	readonly priority = 95;
+	readonly name = 'inline-completions';
+	readonly when = ContextKeyExpr.and(InlineCompletionContextKeys.inlineSuggestionVisible);
+	implementation(accessor: ServicesAccessor) {
+		const accessibleViewService = accessor.get(IAccessibleViewService);
+		const codeEditorService = accessor.get(ICodeEditorService);
+		const show = () => {
+			const editor = codeEditorService.getActiveCodeEditor() || codeEditorService.getFocusedCodeEditor();
+			if (!editor) {
+				return false;
+			}
+			const model = InlineCompletionsController.get(editor)?.model.get();
+			const state = model?.state.get();
+			if (!model || !state) {
+				return false;
+			}
+			const lineText = model.textModel.getLineContent(state.primaryGhostText.lineNumber);
+			const ghostText = state.primaryGhostText.renderForScreenReader(lineText);
+			if (!ghostText) {
+				return false;
+			}
+			const language = editor.getModel()?.getLanguageId() ?? undefined;
+			accessibleViewService.show({
+				id: AccessibleViewProviderId.InlineCompletions,
+				verbositySettingKey: 'accessibility.verbosity.inlineCompletions',
+				provideContent() { return lineText + ghostText; },
+				onClose() {
+					model.stop();
+					editor.focus();
+				},
+				next() {
+					model.next();
+					setTimeout(() => show(), 50);
+				},
+				previous() {
+					model.previous();
+					setTimeout(() => show(), 50);
+				},
+				options: { language, type: this.type }
+			});
+			return true;
+		};
+		return show();
+	}
+	constructor() {
+		super();
+	}
+}
