+++ src/vs/workbench/contrib/accessibility/browser/accessibleViewContributions.ts
@@ -4,15 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Disposable } from 'vs/base/common/lifecycle';
-import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
-import { AccessibilityVerbositySettingId, accessibleViewIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibilityConfiguration';
+import { accessibleViewIsShown } from 'vs/workbench/contrib/accessibility/browser/accessibilityConfiguration';
 import * as strings from 'vs/base/common/strings';
 import { AccessibilityHelpAction, AccessibleViewAction } from 'vs/workbench/contrib/accessibility/browser/accessibleViewActions';
-import { InlineCompletionsController } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionsController';
-import { InlineCompletionContextKeys } from 'vs/editor/contrib/inlineCompletions/browser/inlineCompletionContextKeys';
-import { ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
-import { IAccessibleViewOptions, AccessibleViewType, AccessibleViewProviderId, IAccessibleViewService } from 'vs/platform/accessibility/browser/accessibleView';
+import { AccessibleViewType, IAccessibleViewService } from 'vs/platform/accessibility/browser/accessibleView';
 import { AccessibleViewRegistry } from 'vs/platform/accessibility/browser/accessibleViewRegistry';
 
 export function descriptionForCommand(commandId: string, msg: string, noKbMsg: string, keybindingService: IKeybindingService): string {
@@ -41,7 +37,6 @@ export class AccesibleViewHelpContribution extends Disposable {
 	}
 }
 
-
 export class AccesibleViewContributions extends Disposable {
 	static ID: 'accesibleViewContributions';
 	constructor() {
@@ -53,55 +48,5 @@ export class AccesibleViewContributions extends Disposable {
 				this._register(AccessibilityHelpAction.addImplementation(impl.priority, impl.name, impl.implementation, impl.when));
 			}
 		});
-	}
-}
-
-
-export class InlineCompletionsAccessibleViewContribution extends Disposable {
-	static ID: 'inlineCompletionsAccessibleViewContribution';
-	private _options: IAccessibleViewOptions = { type: AccessibleViewType.View };
-	constructor() {
-		super();
-		this._register(AccessibleViewAction.addImplementation(95, 'inline-completions', accessor => {
-			const accessibleViewService = accessor.get(IAccessibleViewService);
-			const codeEditorService = accessor.get(ICodeEditorService);
-			const show = () => {
-				const editor = codeEditorService.getActiveCodeEditor() || codeEditorService.getFocusedCodeEditor();
-				if (!editor) {
-					return false;
-				}
-				const model = InlineCompletionsController.get(editor)?.model.get();
-				const state = model?.state.get();
-				if (!model || !state) {
-					return false;
-				}
-				const lineText = model.textModel.getLineContent(state.primaryGhostText.lineNumber);
-				const ghostText = state.primaryGhostText.renderForScreenReader(lineText);
-				if (!ghostText) {
-					return false;
-				}
-				this._options.language = editor.getModel()?.getLanguageId() ?? undefined;
-				accessibleViewService.show({
-					id: AccessibleViewProviderId.InlineCompletions,
-					verbositySettingKey: AccessibilityVerbositySettingId.InlineCompletions,
-					provideContent() { return lineText + ghostText; },
-					onClose() {
-						model.stop();
-						editor.focus();
-					},
-					next() {
-						model.next();
-						setTimeout(() => show(), 50);
-					},
-					previous() {
-						model.previous();
-						setTimeout(() => show(), 50);
-					},
-					options: this._options
-				});
-				return true;
-			}; ContextKeyExpr.and(InlineCompletionContextKeys.inlineSuggestionVisible);
-			return show();
-		}));
 	}
 }
